# api/openapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce Website API
  description: |
    A comprehensive ecommerce API that provides functionality for user authentication, 
    product catalog management, shopping cart operations, order processing, and user profile management.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Response Format
    All API responses follow a consistent format:
    - Success responses include `success: true`, optional `message`, `data`, and `timestamp`
    - Error responses include `success: false`, `error` object with `code`, `message`, and optional `details`, and `timestamp`
  version: 1.0.0
  contact:
    name: API Support
    email: support@ecommerce.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.ecommerce.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/auth/admin/login:
    post:
      tags:
        - Authentication
      summary: Admin login
      description: Authenticates an admin using environment credentials and returns JWT tokens
      operationId: adminLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: Admin login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
                          tokens:
                            $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/InvalidCredentialsError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Generates new access and refresh tokens using a valid refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  description: Valid refresh token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          tokens:
                            $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/InvalidTokenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Sends password reset instructions to the user's email
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset instructions sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Resets user password using a valid reset token
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/verify-email:
    get:
      tags:
        - Authentication
      summary: Verify email address
      description: Verifies user email using verification token
      operationId: verifyEmail
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Email verification token
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/resend-verification:
    post:
      tags:
        - Authentication
      summary: Resend email verification
      description: Resends email verification to the authenticated user
      operationId: resendEmailVerification
      responses:
        '200':
          description: Verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/UserNotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Product Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object
          description: Response data (varies by endpoint)
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid request data"
        details:
          type: object
          description: Additional error details

    # Authentication Schemas
...
    AdminLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "admin@ecommerce.com"
        password:
          type: string
          example: "admin123456"
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
          example: "reset-token-here"
        password:
          type: string
          minLength: 8
          example: "newsecurepassword123"
    TokenPair:
      type: object
      required:
        - access_token
        - refresh_token
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    # User Schemas
    User:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isActive
        - emailVerified
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          example: "user@example.com"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        phone:
          type: string
          nullable: true
          example: "+1234567890"
        role:
          type: string
          enum: [customer, admin]
          example: "customer"
        isActive:
          type: boolean
          example: true
        emailVerified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
...
    Address:
      type: object
      required:
        - id
        - userId
        - type
        - firstName
        - lastName
        - address1
        - city
        - state
        - postalCode
        - country
        - isDefault
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        userId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        type:
          type: string
          enum: [shipping, billing]
          example: "shipping"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        company:
          type: string
          nullable: true
          example: "Acme Corp"
        address1:
          type: string
          example: "123 Main St"
        address2:
          type: string
          nullable: true
          example: "Apt 4B"
        city:
          type: string
          example: "New York"
        state:
          type: string
          example: "NY"
        postalCode:
          type: string
          example: "10001"
        country:
          type: string
          example: "US"
        phone:
          type: string
          nullable: true
          example: "+1234567890"
        isDefault:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
...
  responses:
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid request data"
              details: "Email is required"
            timestamp: "2024-01-01T12:00:00Z"
    UnauthorizedError:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
            timestamp: "2024-01-01T12:00:00Z"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
            timestamp: "2024-01-01T12:00:00Z"

    # Specific Error Responses
...
    InvalidCredentialsError:
      description: Invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INVALID_CREDENTIALS"
              message: "Invalid email or password"
            timestamp: "2024-01-01T12:00:00Z"
    InvalidTokenError:
      description: Invalid or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INVALID_TOKEN"
              message: "Invalid or expired token"
            timestamp: "2024-01-01T12:00:00Z"
    UserNotFoundError:
      description: User not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "USER_NOT_FOUND"
              message: "User not found"
            timestamp: "2024-01-01T12:00:00Z"
...